refactor(ast): extract string serialization into strings.ts (#3572 1a)
Move trimQuotes, stringify, jsStringEscape, toRegExpString, stringifyObject,
and getNestedAccessor out of utils/index.ts into a dedicated utils/strings.ts
module, with their tests in strings.test.ts. utils/index.ts re-exports the same
names, so @kubb/ast/utils keeps an identical public surface.
